LayoutTests/inspector/unit-tests/object-expected.txt

Testing WI.Object.


== Running test suite: Object
-- Running test case: WI.Object.prototype.dispatchEventToListeners
Dispatch count: 1

-- Running test case: WI.Object.prototype.awaitEvent
PASS: Await event handler should be dispatched just once.
Dispatch count: 1
